Fix Async sort issue in collection show

+++ app/assets/javascripts/views/collections/show.js
@@ -4,8 +4,7 @@ HealthHunt.Views.CollectionShow = Backbone.CompositeView.extend({
     this.products = this.model.products();
 
     // Set up for composite views
-    this.listenTo(this.products, 'add', this.addProductView);
-    this.products.each(this.addProductView.bind(this));
+    this.listenTo(this.products, 'sync', this.addProductView);
     this.listenTo(this.products, 'remove', this.removeProductView);
   },
 
@@ -20,14 +19,14 @@ HealthHunt.Views.CollectionShow = Backbone.CompositeView.extend({
       collection: this.model
     });
     this.$el.html(content);
-    this.attachSubviews();
+    this.products.each(this.addProductView.bind(this));
     return this;
   },
 
   addProductView: function (product) {
     var subview = new HealthHunt.Views.ProductsIndexItem({
       model: product,
-      collection: this.collection
+      collection: this.products
     });
     this.addSubview('.products', subview);
   },
@@ -38,10 +37,14 @@ HealthHunt.Views.CollectionShow = Backbone.CompositeView.extend({
 
   deleteCollection: function (event) {
     event.preventDefault();
-    this.model.destroy({
-      success: function () {
-        Backbone.history.navigate("#/collections", { trigger: true });
-      }
-    });
+    var ask = confirm("Are you sure you want to delete this collection?");
+
+    if(ask){
+      this.model.destroy({
+        success: function () {
+          Backbone.history.navigate("#/collections", { trigger: true });
+        }
+      });
+     }
   }
 });
